requirements.txt
paho-mqtt==2.2.1
boto3==1.26.0
Flask==2.2.3